Commiting the solution I came up with the interview. 1 known issue.
0 0 1
0 1 1
0 0 1

In this situation, the application would count 2 islands. I'm currently
only checking the above and left for islands, when in this edge case it
would need to have checked to the right as well to be accurate. The
right would have only been added to the visited list if we checked
below. In short, I need to check in all directions.